Use Base64#strict_encoding64 to avoid newlines being added
The problem was that document_number was being encoded as "MjYzMTgzNDRM\n" (example),
but the RegistrationForm received "MjYzMTgzNDRM\r\n". This caused the User to be saved
with the wrong String and then IdentityDocumentForm was not being able to validate that
no other User in the Organization was registered with the same encoded data.